+++ src/main/java/arozxmod/BoatFlying.java
@@ -5,11 +5,16 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class BoatFlying {
+    public static final Logger LOGGER = LoggerFactory.getLogger("BoatFlyingMod");
+
     private static boolean changeForwardSpeed;
     private static double upwardSpeed = 0.15;
     private static double forwardSpeed;
+    static int toggle = 0;
 
     public static void onUpdate() {
         // Check if the player is riding a vehicle
@@ -26,6 +31,12 @@ public class BoatFlying {
         // Up/down movement
         if (MinecraftClient.getInstance().options.jumpKey.isPressed()) {
             motionY = upwardSpeed;
+            if (toggle == 0 || velocity.y <= -0.035) {
+                toggle = 40;
+                // debug
+                LOGGER.info("Fall"+toggle);
+            }
+            toggle--;
         } else if (MinecraftClient.getInstance().options.sprintKey.isPressed()) {
             motionY = velocity.y;
         }