+++ src/main/java/arozxmod/screens/AutoclickerScreen.java
@@ -1,13 +1,12 @@
 package arozxmod.screens;
 
 import arozxmod.Config;
-import arozxmod.Autoclicker;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.ButtonWidget;
+import net.minecraft.client.gui.widget.SliderWidget;
 import net.minecraft.client.option.GameOptions;
-import net.minecraft.screen.ScreenTexts;
 import net.minecraft.text.Text;
 
 @Environment(EnvType.CLIENT)
@@ -40,27 +39,57 @@ public class AutoclickerScreen extends Screen {
 
     @Override
     protected void init() {
+        // CPS Slider
+        this.addDrawableChild(new SliderWidget(this.width / 2 - 102, this.height / 6 + 58, 204, 20, Text.of("CPS"), 0) {
+            @Override
+            protected void updateMessage() {
+                this.setMessage(Text.of("CPS: " + (int) (this.value*10)));
+            }
+
+            @Override
+            protected void applyValue() {
+                Config.cps = (int) (this.value*10);
+            }
+        });
         // Set cps button
-        this.addDrawableChild(new ButtonWidget(this.width / 2 - 150, this.height / 6 + 118, 150, 20,
-                cpsText(), (button) -> {
-            Autoclicker.cps =5;
-            Config.cpsEnabled = !Config.cpsEnabled;
-            button.setMessage(cpsText());
-        }));
+        this.addDrawableChild(
+                ButtonWidget.builder(
+                        cpsText(), button -> {
+                            Config.cpsEnabled = !Config.cpsEnabled;
+                            button.setMessage(cpsText());
+                        }
+                        )
+                        .dimensions(this.width / 2 - 102, this.height / 6 + 80, 204, 20)
+                        .build());
+
         // Right autoclicker button
-        this.addDrawableChild(new ButtonWidget(this.width / 2, this.height / 6 + 146, 150, 20,
-                autoclickerTextR(), (button) -> {
-            Config.autoclickerEnabledR = !Config.autoclickerEnabledR;
-            button.setMessage(autoclickerTextR());
-        }));
+        this.addDrawableChild(
+                ButtonWidget.builder(
+                        autoclickerTextR(), button -> {
+                            Config.autoclickerEnabledR = !Config.autoclickerEnabledR;
+                            button.setMessage(autoclickerTextR());
+                        }
+                        )
+                        .dimensions(this.width / 2 - 102, this.height / 6 + 108, 204, 20)
+                        .build());
+
         // Left autoclicker button
-        this.addDrawableChild(new ButtonWidget(this.width / 2 - 150, this.height / 6 + 146, 150, 20,
-                autoclickerTextL(), (button) -> {
-            Config.autoclickerEnabledL = !Config.autoclickerEnabledL;
-            button.setMessage(autoclickerTextL());
-        }));
+        this.addDrawableChild(
+                ButtonWidget.builder(
+                        autoclickerTextL(), button -> {
+                            Config.autoclickerEnabledL = !Config.autoclickerEnabledL;
+                            button.setMessage(autoclickerTextL());
+                        }
+                        )
+                        .dimensions(this.width / 2 - 102, this.height / 6 + 136, 204, 20)
+                        .build());
+
         // Back button
-        this.addDrawableChild(new ButtonWidget(this.width / 2 - 100, this.height / 6 + 174, 200, 20,
-                ScreenTexts.BACK, (button) -> this.client.setScreen(this.parent)));
+        this.addDrawableChild(
+                ButtonWidget.builder(
+                        Text.of("Back"), button -> this.client.setScreen(this.parent)
+                        )
+                        .dimensions(this.width / 2 - 102, this.height / 6 + 174, 204, 20)
+                        .build());
     }
 }