+++ Changelog.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased]
+## [1.0.2] - 2023-08-07
 
 ### Added
 
-- New menu
+- Slider for autoclicker value
+- Re-written for 1.20.1
 
 ### Changed